cloud_optimize_env/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-optimize-env"
version = "0.1.0"
description = "Cloud cost optimization RL environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[project.scripts]
server = "cloud_optimize_env.server.app:main"

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]
